+++ Combo.h
@@ -23,6 +23,7 @@
 #define COMBO_H
 
 #include "Graph.h"
+#include <cstdint>
 #include <optional>
 #include <random>
 #include <vector>
@@ -30,7 +31,7 @@
 class ComboAlgorithm {
 public:
     ComboAlgorithm();
-    explicit ComboAlgorithm(std::optional<unsigned long long> random_seed, int num_split_attempts, int fixed_split_step);
+    explicit ComboAlgorithm(std::optional<uint_fast32_t> random_seed, int num_split_attempts, int fixed_split_step);
     ComboAlgorithm(int num_split_attempts, int fixed_split_step);
     void Run(Graph& graph, std::optional<size_t> max_communities = std::nullopt, bool start_separate = false);
     void SetFixedSplitStep(int fixed_split_step) {m_fixed_split_step = fixed_split_step;}
